+++ src/ayatana-indicator-0.4.pc.in.in
