+++ src/dbus-shared.h
