+++ src/indicator-desktop-shortcuts.c
