+++ src/indicator-image-helper.c
