+++ src/indicator-image-helper.h
