+++ src/indicator-object-enum-types.c.template
