+++ src/indicator-object-enum-types.h.template
