+++ src/indicator-object.c
