+++ src/indicator-object.h
