+++ src/indicator-service-manager.c
