+++ src/indicator-service-manager.h
