+++ src/indicator-service.c
