+++ tests/Makefile.am
@@ -9,7 +9,7 @@ TESTS =
 DISTCLEANFILES =
 XFAIL_TESTS =
 
-check_PROGRAMS = 
+check_PROGRAMS =
 
 lib_LTLIBRARIES = \
 	libdummy-indicator-blank.la \
@@ -46,7 +46,7 @@ test_loader_CFLAGS = \
 
 test_loader_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 #############################
@@ -66,7 +66,7 @@ test_desktop_shortcuts_CFLAGS = \
 
 test_desktop_shortcuts_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 DS_XML_REPORT  = desktop-shortcuts-check-results.xml
@@ -99,7 +99,7 @@ libdummy_indicator_blank_la_CFLAGS = \
 
 libdummy_indicator_blank_la_LIBADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 libdummy_indicator_blank_la_LDFLAGS = \
@@ -120,7 +120,7 @@ libdummy_indicator_null_la_CFLAGS = \
 
 libdummy_indicator_null_la_LIBADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 libdummy_indicator_null_la_LDFLAGS = \
@@ -141,7 +141,7 @@ libdummy_indicator_signaler_la_CFLAGS = \
 
 libdummy_indicator_signaler_la_LIBADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 libdummy_indicator_signaler_la_LDFLAGS = \
@@ -162,7 +162,7 @@ libdummy_indicator_simple_la_CFLAGS = \
 
 libdummy_indicator_simple_la_LIBADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 libdummy_indicator_simple_la_LDFLAGS = \
@@ -184,7 +184,7 @@ libdummy_indicator_entry_func_la_CFLAGS = \
 
 libdummy_indicator_entry_func_la_LIBADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 libdummy_indicator_entry_func_la_LDFLAGS = \
@@ -205,7 +205,7 @@ libdummy_indicator_visible_la_CFLAGS = \
 
 libdummy_indicator_visible_la_LIBADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 libdummy_indicator_visible_la_LDFLAGS = \
@@ -227,7 +227,7 @@ service_shutdown_timeout_CFLAGS = \
 
 service_shutdown_timeout_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 service-shutdown-timeout-tester: service-shutdown-timeout Makefile
@@ -253,7 +253,7 @@ service_manager_no_connect_CFLAGS = \
 
 service_manager_no_connect_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 service-manager-no-connect-tester: service-manager-no-connect Makefile.am
@@ -285,7 +285,7 @@ service_manager_connect_CFLAGS = \
 
 service_manager_connect_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 check_PROGRAMS += service-manager-connect-service
@@ -299,7 +299,7 @@ service_manager_connect_service_CFLAGS = \
 
 service_manager_connect_service_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 service-manager-connect-tester: service-manager-connect service-manager-connect-service session.conf service-manager-connect.service Makefile.am
@@ -332,7 +332,7 @@ service_version_manager_CFLAGS = \
 
 service_version_manager_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 check_PROGRAMS += service-version-bad-service
@@ -347,7 +347,7 @@ service_version_bad_service_CFLAGS = \
 
 service_version_bad_service_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 check_PROGRAMS += service-version-good-service
@@ -362,7 +362,7 @@ service_version_good_service_CFLAGS = \
 
 service_version_good_service_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 service-version-tester: service-version-manager service-version-bad-service service-version-good-service session.conf service-version-bad.service service-version-good.service Makefile.am
@@ -389,7 +389,7 @@ service_version_multiwatch_manager_CFLAGS = \
 
 service_version_multiwatch_manager_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 check_PROGRAMS += service-version-multiwatch-manager-impolite
@@ -404,7 +404,7 @@ service_version_multiwatch_manager_impolite_CFLAGS = \
 
 service_version_multiwatch_manager_impolite_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 check_PROGRAMS += service-version-multiwatch-service
@@ -419,7 +419,7 @@ service_version_multiwatch_service_CFLAGS = \
 
 service_version_multiwatch_service_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 service-version-multiwatch-tester: service-version-multiwatch-manager service-version-multiwatch-service service-version-multiwatch-manager-impolite Makefile.am
@@ -445,7 +445,7 @@ service_manager_nostart_connect_CFLAGS = \
 
 service_manager_nostart_connect_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
 
 service-manager-connect-nostart-tester: service-manager-nostart-connect service-manager-connect-service Makefile.am
@@ -496,7 +496,7 @@ test_indicator_ng_CFLAGS = \
 
 test_indicator_ng_LDADD = \
 	$(LIBINDICATOR_LIBS) \
-	-L$(top_builddir)/libayatana-indicator/.libs \
+	-L$(top_builddir)/src/.libs \
 	$(INDICATOR_LIB)
