+++ tests/dummy-indicator-null.c
@@ -24,8 +24,8 @@ License along with this library. If not, see
 #include <glib.h>
 #include <glib-object.h>
 
-#include "libayatana-indicator/indicator.h"
-#include "libayatana-indicator/indicator-object.h"
+#include "src/indicator.h"
+#include "src/indicator-object.h"
 
 #define DUMMY_INDICATOR_NULL_TYPE            (dummy_indicator_null_get_type ())
 #define DUMMY_INDICATOR_NULL(obj)            (G_TYPE_CHECK_INSTANCE_CAST ((obj), DUMMY_INDICATOR_NULL_TYPE, DummyIndicatorNull))
@@ -38,11 +38,11 @@ typedef struct _DummyIndicatorNull      DummyIndicatorNull;
 typedef struct _DummyIndicatorNullClass DummyIndicatorNullClass;
 
 struct _DummyIndicatorNullClass {
-	IndicatorObjectClass parent_class;
+    IndicatorObjectClass parent_class;
 };
 
 struct _DummyIndicatorNull {
-	IndicatorObject parent;
+    IndicatorObject parent;
 };
 
 GType dummy_indicator_null_get_type (void);
@@ -54,24 +54,24 @@ INDICATOR_SET_TYPE(DUMMY_INDICATOR_NULL_TYPE)
 GtkLabel *
 get_label (IndicatorObject * io)
 {
-	return NULL;
+    return NULL;
 }
 
 GtkImage *
 get_icon (IndicatorObject * io)
 {
-	return NULL;
+    return NULL;
 }
 
 GtkMenu *
 get_menu (IndicatorObject * io)
 {
-	return NULL;
+    return NULL;
 }
 const gchar *
 get_accessible_desc (IndicatorObject * io)
 {
-	return NULL;
+    return NULL;
 }
 
 static void dummy_indicator_null_class_init (DummyIndicatorNullClass *klass);
@@ -84,40 +84,40 @@ G_DEFINE_TYPE (DummyIndicatorNull, dummy_indicator_null, INDICATOR_OBJECT_TYPE);
 static void
 dummy_indicator_null_class_init (DummyIndicatorNullClass *klass)
 {
-	GObjectClass *object_class = G_OBJECT_CLASS (klass);
+    GObjectClass *object_class = G_OBJECT_CLASS (klass);
 
-	object_class->dispose = dummy_indicator_null_dispose;
-	object_class->finalize = dummy_indicator_null_finalize;
+    object_class->dispose = dummy_indicator_null_dispose;
+    object_class->finalize = dummy_indicator_null_finalize;
 
-	IndicatorObjectClass * io_class = INDICATOR_OBJECT_CLASS(klass);
+    IndicatorObjectClass * io_class = INDICATOR_OBJECT_CLASS(klass);
 
-	io_class->get_label = get_label;
-	io_class->get_image = get_icon;
-	io_class->get_menu = get_menu;
-	io_class->get_accessible_desc = get_accessible_desc;
+    io_class->get_label = get_label;
+    io_class->get_image = get_icon;
+    io_class->get_menu = get_menu;
+    io_class->get_accessible_desc = get_accessible_desc;
 
-	return;
+    return;
 }
 
 static void
 dummy_indicator_null_init (DummyIndicatorNull *self)
 {
 
-	return;
+    return;
 }
 
 static void
 dummy_indicator_null_dispose (GObject *object)
 {
 
-	G_OBJECT_CLASS (dummy_indicator_null_parent_class)->dispose (object);
-	return;
+    G_OBJECT_CLASS (dummy_indicator_null_parent_class)->dispose (object);
+    return;
 }
 
 static void
 dummy_indicator_null_finalize (GObject *object)
 {
 
-	G_OBJECT_CLASS (dummy_indicator_null_parent_class)->finalize (object);
-	return;
+    G_OBJECT_CLASS (dummy_indicator_null_parent_class)->finalize (object);
+    return;
 }
