+++ tests/dummy-indicator-simple.c
@@ -23,8 +23,8 @@ License along with this library. If not, see
 #include <glib.h>
 #include <glib-object.h>
 
-#include "libayatana-indicator/indicator.h"
-#include "libayatana-indicator/indicator-object.h"
+#include "src/indicator.h"
+#include "src/indicator-object.h"
 
 #define DUMMY_INDICATOR_SIMPLE_TYPE            (dummy_indicator_simple_get_type ())
 #define DUMMY_INDICATOR_SIMPLE(obj)            (G_TYPE_CHECK_INSTANCE_CAST ((obj), DUMMY_INDICATOR_SIMPLE_TYPE, DummyIndicatorSimple))
@@ -37,11 +37,11 @@ typedef struct _DummyIndicatorSimple      DummyIndicatorSimple;
 typedef struct _DummyIndicatorSimpleClass DummyIndicatorSimpleClass;
 
 struct _DummyIndicatorSimpleClass {
-	IndicatorObjectClass parent_class;
+    IndicatorObjectClass parent_class;
 };
 
 struct _DummyIndicatorSimple {
-	IndicatorObject parent;
+    IndicatorObject parent;
 };
 
 GType dummy_indicator_simple_get_type (void);
@@ -52,30 +52,30 @@ INDICATOR_SET_TYPE(DUMMY_INDICATOR_SIMPLE_TYPE)
 GtkLabel *
 get_label (IndicatorObject * io)
 {
-	return GTK_LABEL(gtk_label_new("Simple Item"));
+    return GTK_LABEL(gtk_label_new("Simple Item"));
 }
 
 GtkImage *
 get_icon (IndicatorObject * io)
 {
-	return GTK_IMAGE(gtk_image_new());
+    return GTK_IMAGE(gtk_image_new());
 }
 
 GtkMenu *
 get_menu (IndicatorObject * io)
 {
-	GtkMenu * main_menu = GTK_MENU(gtk_menu_new());
-	GtkWidget * loading_item = gtk_menu_item_new_with_label("Loading...");
-	gtk_menu_shell_append(GTK_MENU_SHELL(main_menu), loading_item);
-	gtk_widget_show(GTK_WIDGET(loading_item));
+    GtkMenu * main_menu = GTK_MENU(gtk_menu_new());
+    GtkWidget * loading_item = gtk_menu_item_new_with_label("Loading...");
+    gtk_menu_shell_append(GTK_MENU_SHELL(main_menu), loading_item);
+    gtk_widget_show(GTK_WIDGET(loading_item));
 
-	return main_menu;
+    return main_menu;
 }
 
 const gchar *
 get_accessible_desc (IndicatorObject * io)
 {
-	return "Simple Item";
+    return "Simple Item";
 }
 
 static void dummy_indicator_simple_class_init (DummyIndicatorSimpleClass *klass);
@@ -88,40 +88,40 @@ G_DEFINE_TYPE (DummyIndicatorSimple, dummy_indicator_simple, INDICATOR_OBJECT_TY
 static void
 dummy_indicator_simple_class_init (DummyIndicatorSimpleClass *klass)
 {
-	GObjectClass *object_class = G_OBJECT_CLASS (klass);
+    GObjectClass *object_class = G_OBJECT_CLASS (klass);
 
-	object_class->dispose = dummy_indicator_simple_dispose;
-	object_class->finalize = dummy_indicator_simple_finalize;
+    object_class->dispose = dummy_indicator_simple_dispose;
+    object_class->finalize = dummy_indicator_simple_finalize;
 
-	IndicatorObjectClass * io_class = INDICATOR_OBJECT_CLASS(klass);
+    IndicatorObjectClass * io_class = INDICATOR_OBJECT_CLASS(klass);
 
-	io_class->get_label = get_label;
-	io_class->get_image = get_icon;
-	io_class->get_menu = get_menu;
-	io_class->get_accessible_desc = get_accessible_desc;
+    io_class->get_label = get_label;
+    io_class->get_image = get_icon;
+    io_class->get_menu = get_menu;
+    io_class->get_accessible_desc = get_accessible_desc;
 
-	return;
+    return;
 }
 
 static void
 dummy_indicator_simple_init (DummyIndicatorSimple *self)
 {
 
-	return;
+    return;
 }
 
 static void
 dummy_indicator_simple_dispose (GObject *object)
 {
 
-	G_OBJECT_CLASS (dummy_indicator_simple_parent_class)->dispose (object);
-	return;
+    G_OBJECT_CLASS (dummy_indicator_simple_parent_class)->dispose (object);
+    return;
 }
 
 static void
 dummy_indicator_simple_finalize (GObject *object)
 {
 
-	G_OBJECT_CLASS (dummy_indicator_simple_parent_class)->finalize (object);
-	return;
+    G_OBJECT_CLASS (dummy_indicator_simple_parent_class)->finalize (object);
+    return;
 }
