+++ tests/service-shutdown-timeout.c
@@ -22,7 +22,7 @@ License along with this library. If not, see
 
 
 #include <glib.h>
-#include "libayatana-indicator/indicator-service.h"
+#include "src/indicator-service.h"
 
 static GMainLoop * mainloop = NULL;
 static gboolean passed = FALSE;
