+++ tests/test-indicator-ng.c
@@ -1,5 +1,5 @@
 
-#include <libayatana-indicator/indicator-ng.h>
+#include "src/indicator-ng.h"
 
 static void
 indicator_ng_test_func (gconstpointer user_data)
